Target and Training Datasets
I did not include feature engineering using dates in order not to complicate too much. Every initial information exist in these new datasets. I consider the problem as predicting both the date when selling occurs and the units if selling occurs. Therefore, in the target_datasets two columns "sold_or_not" and "units" are all NA; waiting to be predicted.